requirements.txt
PyYAML==5.4.1
whois-api==1.1.0
